Get it pip installing and running native + python code, albeit weirdly
It currently pip installs the native library (which *must* match
the name of the cpp header) *next to* the python module. This
means that to call native functions, you have to import _dig. Eg:

>>> import _dig
>>> _dig.swigTest(4)
swigTest: received int: 4
5
>>>

and to call pure python functions you have to do:

>>> from dig import tmp
>>> tmp.foo()
hello
>>>

It would be nice to have the native code + wrappers in the normal
python module, so it could be, eg, "from dig import native", but
setuptools seems bent on putting the native-derived files in the
parent directory.