FIX Scons build
Sconscripts in src/ and test/ weren't generating the right relative paths
because we were appending the child directory's basenames to the current
directory, rather than to their parent directories. This resulted in adding
dirs that didn't exist to the include path instead of the real ones for the
eigen subdirectories (e.g., src/Cholesky, instead of
src/external/eigen/src/Cholesky). Also note that the current SConstructs are
including these files in the src/ dir, not in the build/ dir, since they're
not build targets and thus aren't copied to the build/ dir.