FIX bugs in subseq local_optima
We had candidateVal declared as a length_t, which was causing it to truncate
doubles to integers before making comparisons; this caused smaller floats to
erroneously compare as being equal.

Additional fix: no longer reports an index as a relative optimum if it happens
to be the first index after a previous optimum is flushed and no further optima
appear within minSpacing in order to replace it.

Also, changed it to select later values in the case of equality.