get kmeans index ported to using IndexConfig
Segfaulting was because we had the designated ctor calls for the
other indexes in the IndexConfig-based ctor bodies, not the
initializer lists. Consequently, it was default constructing
the inner index, then constructing another one, then immediately
freeing it. That was pretty nasty.

Note that L2KmeansIndex still hasn't been built/tested with another KmeansIndex as the inner index.